# 8263006: Add optimization for Max(*)Node and Min(*)Node