--- a/src/hotspot/share/opto/addnode.cpp
+++ b/src/hotspot/share/opto/addnode.cpp
@@ -78,7 +78,7 @@ static bool commute(PhaseGVN* phase, Node* add) {
     if ((in11 == in21 && in12 == in22) ||
         (in11 == in22 && in12 == in21)) {
       add->set_req(1, in11);
-      add->set_req(2, in22);
+      add->set_req(2, in12);
       PhaseIterGVN* igvn = phase->is_IterGVN();
       if (igvn) {
         igvn->_worklist.push(in1);