--- a/src/hotspot/share/opto/addnode.hpp
+++ b/src/hotspot/share/opto/addnode.hpp
@@ -68,6 +68,12 @@ public:
   // Supplied function to return the additive identity type
   virtual const Type *add_id() const = 0;
 
+  // Supplied function to return the additive opcode
+  virtual int max_opcode() const = 0;
+
+  // Supplied function to return the multiplicative opcode
+  virtual int min_opcode() const = 0;
+
   virtual bool operates_on(BasicType bt, bool signed_int) const {
     assert(bt == T_INT || bt == T_LONG, "unsupported");
     return false;
@@ -84,6 +90,8 @@ public:
   virtual const Type *add_ring( const Type *, const Type * ) const;
   virtual const Type *add_id() const { return TypeInt::ZERO; }
   virtual const Type *bottom_type() const { return TypeInt::INT; }
+  int max_opcode() const { return Op_MaxI; }
+  int min_opcode() const { return Op_MinI; }
   virtual Node *Ideal(PhaseGVN *phase, bool can_reshape);
   virtual Node* Identity(PhaseGVN* phase);
   virtual bool operates_on(BasicType bt, bool signed_int) const {
@@ -102,6 +110,8 @@ public:
   virtual const Type *add_ring( const Type *, const Type * ) const;
   virtual const Type *add_id() const { return TypeLong::ZERO; }
   virtual const Type *bottom_type() const { return TypeLong::LONG; }
+  int max_opcode() const { return Op_MaxL; }
+  int min_opcode() const { return Op_MinL; }
   virtual Node *Ideal(PhaseGVN *phase, bool can_reshape);
   virtual Node* Identity(PhaseGVN* phase);
   virtual bool operates_on(BasicType bt, bool signed_int) const {
@@ -122,6 +132,8 @@ public:
   virtual const Type *add_ring( const Type *, const Type * ) const;
   virtual const Type *add_id() const { return TypeF::ZERO; }
   virtual const Type *bottom_type() const { return Type::FLOAT; }
+  int max_opcode() const { return Op_MaxF; }
+  int min_opcode() const { return Op_MinF; }
   virtual Node* Identity(PhaseGVN* phase) { return this; }
   virtual uint ideal_reg() const { return Op_RegF; }
 };
@@ -137,6 +149,8 @@ public:
   virtual const Type *add_ring( const Type *, const Type * ) const;
   virtual const Type *add_id() const { return TypeD::ZERO; }
   virtual const Type *bottom_type() const { return Type::DOUBLE; }
+  int max_opcode() const { return Op_MaxD; }
+  int min_opcode() const { return Op_MinD; }
   virtual Node* Identity(PhaseGVN* phase) { return this; }
   virtual uint ideal_reg() const { return Op_RegD; }
 };
@@ -183,6 +197,8 @@ public:
   virtual const Type *add_ring( const Type *, const Type * ) const;
   virtual const Type *add_id() const { return TypeInt::ZERO; }
   virtual const Type *bottom_type() const { return TypeInt::INT; }
+  int max_opcode() const { return Op_MaxI; }
+  int min_opcode() const { return Op_MinI; }
   virtual Node* Identity(PhaseGVN* phase);
   virtual uint ideal_reg() const { return Op_RegI; }
   virtual Node *Ideal(PhaseGVN *phase, bool can_reshape);
@@ -198,6 +214,8 @@ public:
   virtual const Type *add_ring( const Type *, const Type * ) const;
   virtual const Type *add_id() const { return TypeLong::ZERO; }
   virtual const Type *bottom_type() const { return TypeLong::LONG; }
+  int max_opcode() const { return Op_MaxL; }
+  int min_opcode() const { return Op_MinL; }
   virtual Node* Identity(PhaseGVN* phase);
   virtual uint ideal_reg() const { return Op_RegL; }
   virtual Node *Ideal(PhaseGVN *phase, bool can_reshape);
@@ -212,6 +230,8 @@ public:
   virtual const Type *add_ring( const Type *, const Type * ) const;
   virtual const Type *add_id() const { return TypeInt::ZERO; }
   virtual const Type *bottom_type() const { return TypeInt::INT; }
+  int max_opcode() const { return Op_MaxI; }
+  int min_opcode() const { return Op_MinI; }
   virtual const Type *Value(PhaseGVN *phase) const;
   virtual uint ideal_reg() const { return Op_RegI; }
 };
@@ -225,6 +245,8 @@ public:
   virtual const Type *add_ring( const Type *, const Type * ) const;
   virtual const Type *add_id() const { return TypeLong::ZERO; }
   virtual const Type *bottom_type() const { return TypeLong::LONG; }
+  int max_opcode() const { return Op_MaxL; }
+  int min_opcode() const { return Op_MinL; }
   virtual const Type *Value(PhaseGVN *phase) const;
   virtual uint ideal_reg() const { return Op_RegL; }
 };
@@ -241,6 +263,8 @@ private:
 public:
   MaxNode( Node *in1, Node *in2 ) : AddNode(in1,in2) {}
   virtual int Opcode() const = 0;
+  virtual int max_opcode() const = 0;
+  virtual int min_opcode() const = 0;
 
   static Node* unsigned_max(Node* a, Node* b, const Type* t, PhaseGVN& gvn) {
     return build_min_max(a, b, true, true, t, gvn);
@@ -280,6 +304,8 @@ public:
   virtual const Type *add_id() const { return TypeInt::make(min_jint); }
   virtual const Type *bottom_type() const { return TypeInt::INT; }
   virtual uint ideal_reg() const { return Op_RegI; }
+  int max_opcode() const { return Op_MaxI; }
+  int min_opcode() const { return Op_MinI; }
 };
 
 //------------------------------MinINode---------------------------------------
@@ -293,6 +319,8 @@ public:
   virtual const Type *add_id() const { return TypeInt::make(max_jint); }
   virtual const Type *bottom_type() const { return TypeInt::INT; }
   virtual uint ideal_reg() const { return Op_RegI; }
+  int max_opcode() const { return Op_MaxI; }
+  int min_opcode() const { return Op_MinI; }
   virtual Node *Ideal(PhaseGVN *phase, bool can_reshape);
 };
 
@@ -306,6 +334,8 @@ public:
   virtual const Type *add_id() const { return TypeLong::make(min_jlong); }
   virtual const Type *bottom_type() const { return TypeLong::LONG; }
   virtual uint ideal_reg() const { return Op_RegL; }
+  int max_opcode() const { return Op_MaxL; }
+  int min_opcode() const { return Op_MinL; }
 };
 
 //------------------------------MinLNode---------------------------------------
@@ -318,6 +348,8 @@ public:
   virtual const Type *add_id() const { return TypeLong::make(max_jlong); }
   virtual const Type *bottom_type() const { return TypeLong::LONG; }
   virtual uint ideal_reg() const { return Op_RegL; }
+  int max_opcode() const { return Op_MaxL; }
+  int min_opcode() const { return Op_MinL; }
 };
 
 //------------------------------MaxFNode---------------------------------------
@@ -330,6 +362,8 @@ public:
   virtual const Type *add_id() const { return TypeF::NEG_INF; }
   virtual const Type *bottom_type() const { return Type::FLOAT; }
   virtual uint ideal_reg() const { return Op_RegF; }
+  int max_opcode() const { return Op_MaxF; }
+  int min_opcode() const { return Op_MinF; }
 };
 
 //------------------------------MinFNode---------------------------------------
@@ -342,6 +376,8 @@ public:
   virtual const Type *add_id() const { return TypeF::POS_INF; }
   virtual const Type *bottom_type() const { return Type::FLOAT; }
   virtual uint ideal_reg() const { return Op_RegF; }
+  int max_opcode() const { return Op_MaxF; }
+  int min_opcode() const { return Op_MinF; }
 };
 
 //------------------------------MaxDNode---------------------------------------
@@ -354,6 +390,8 @@ public:
   virtual const Type *add_id() const { return TypeD::NEG_INF; }
   virtual const Type *bottom_type() const { return Type::DOUBLE; }
   virtual uint ideal_reg() const { return Op_RegD; }
+  int max_opcode() const { return Op_MaxD; }
+  int min_opcode() const { return Op_MinD; }
 };
 
 //------------------------------MinDNode---------------------------------------
@@ -366,6 +404,8 @@ public:
   virtual const Type *add_id() const { return TypeD::POS_INF; }
   virtual const Type *bottom_type() const { return Type::DOUBLE; }
   virtual uint ideal_reg() const { return Op_RegD; }
+  int max_opcode() const { return Op_MaxD; }
+  int min_opcode() const { return Op_MinD; }
 };
 
 #endif // SHARE_OPTO_ADDNODE_HPP