--- a/src/hotspot/share/opto/mulnode.cpp
+++ b/src/hotspot/share/opto/mulnode.cpp
@@ -74,8 +74,8 @@ Node *MulNode::Ideal(PhaseGVN *phase, bool can_reshape) {
     Node *in21 = in(2)->in(1);
     Node *in22 = in(2)->in(2);
 
-    if (in11 == in21 && in12 == in22 ||
-        in11 == in22 && in12 == in21) {
+    if ((in11 == in21 && in12 == in22) ||
+        (in11 == in22 && in12 == in21)) {
       set_req(1, in11);
       set_req(2, in22);
       PhaseIterGVN* igvn = phase->is_IterGVN();