--- a/src/hotspot/share/opto/mulnode.hpp
+++ b/src/hotspot/share/opto/mulnode.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1997, 2020, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1997, 2021, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -75,6 +75,11 @@ public:
   // Supplied function to return the multiplicative opcode
   virtual int mul_opcode() const = 0;
 
+  // Supplied function to return the additive opcode
+  virtual int max_opcode() const = 0;
+
+  // Supplied function to return the multiplicative opcode
+  virtual int min_opcode() const = 0;
 };
 
 //------------------------------MulINode---------------------------------------
@@ -89,6 +94,8 @@ public:
   const Type *add_id() const { return TypeInt::ZERO; }
   int add_opcode() const { return Op_AddI; }
   int mul_opcode() const { return Op_MulI; }
+  int max_opcode() const { return Op_MaxI; }
+  int min_opcode() const { return Op_MinI; }
   const Type *bottom_type() const { return TypeInt::INT; }
   virtual uint ideal_reg() const { return Op_RegI; }
 };
@@ -105,6 +112,8 @@ public:
   const Type *add_id() const { return TypeLong::ZERO; }
   int add_opcode() const { return Op_AddL; }
   int mul_opcode() const { return Op_MulL; }
+  int max_opcode() const { return Op_MaxL; }
+  int min_opcode() const { return Op_MinL; }
   const Type *bottom_type() const { return TypeLong::LONG; }
   virtual uint ideal_reg() const { return Op_RegL; }
 };
@@ -121,6 +130,8 @@ public:
   const Type *add_id() const { return TypeF::ZERO; }
   int add_opcode() const { return Op_AddF; }
   int mul_opcode() const { return Op_MulF; }
+  int max_opcode() const { return Op_MaxF; }
+  int min_opcode() const { return Op_MinF; }
   const Type *bottom_type() const { return Type::FLOAT; }
   virtual uint ideal_reg() const { return Op_RegF; }
 };
@@ -136,6 +147,8 @@ public:
   const Type *add_id() const { return TypeD::ZERO; }
   int add_opcode() const { return Op_AddD; }
   int mul_opcode() const { return Op_MulD; }
+  int max_opcode() const { return Op_MaxD; }
+  int min_opcode() const { return Op_MinD; }
   const Type *bottom_type() const { return Type::DOUBLE; }
   virtual uint ideal_reg() const { return Op_RegD; }
 };
@@ -165,6 +178,8 @@ public:
   const Type *add_id() const { return TypeInt::ZERO; }
   int add_opcode() const { return Op_OrI; }
   int mul_opcode() const { return Op_AndI; }
+  int max_opcode() const { return Op_MaxI; }
+  int min_opcode() const { return Op_MinI; }
   virtual uint ideal_reg() const { return Op_RegI; }
 };
 
@@ -182,6 +197,8 @@ public:
   const Type *add_id() const { return TypeLong::ZERO; }
   int add_opcode() const { return Op_OrL; }
   int mul_opcode() const { return Op_AndL; }
+  int max_opcode() const { return Op_MaxL; }
+  int min_opcode() const { return Op_MinL; }
   virtual uint ideal_reg() const { return Op_RegL; }
 };