… test/micro/org/openjdk/bench/vm/compiler/MaxMinOptimizeTest.java