@vnkozlov 2021-04-17
Do you have a real example in Java applications which benefit from this optimization?
We should not add and **support** code which would never be used in real world.

Optimization will not work for Integer because of `_min` and `_max` intrinsic which generates `cmove`:
https://github.com/openjdk/jdk/blob/master/src/hotspot/share/opto/library_call.cpp#L1806

I am not sure if this optimization will always work for float/double because of NaN values.

You need to verify results for all edge cases.