@Wanghuang-Huawei 2021-04-17
Thank you for your review.
> Do you have a real example in Java applications which benefit from this optimization?
> We should not add and **support** code which would never be used in real world.
> 
Yes. We refined this optimization from our internal software experience. For instance, the model `min( max(a,b) , min(a,b)))` exists in many source codes in especial some AI project.
> Optimization will not work for Integer because of `_min` and `_max` intrinsic which generates `cmove`:
> https://github.com/openjdk/jdk/blob/master/src/hotspot/share/opto/library_call.cpp#L1806
> 
Yes. Adding `MaxINode`'s `max_opcode`  is just for `max_opcode` method is abstract. Our test cases is for float types only. 
> I am not sure if this optimization will always work for float/double because of NaN values.
> 
> You need to verify results for all edge cases.
I have tested that and showed in my comments. The test cases for NaN values and other special values are listed here 
```java
import java.lang.Math;

public class Test {

    public static void main(String[] args) throws Exception {
      Test m = new Test();
      m.test();
    }

    public void test() throws Exception {
      double[] num = new double[9];
      num[0] = 1; num[1] = 0; num[2] = -0;
      num[3] = Double.POSITIVE_INFINITY; 
      num[4] = Double.NEGATIVE_INFINITY;
      num[5] = Double.NaN;
      num[6] = Double.MAX_VALUE;
      num[7] = Double.MIN_VALUE;
      num[8] = Double.MIN_NORMAL;

      for(int i = 0; i < 9; i++) {
        for(int j = 0; j < 9; j++) {
          check(add_opt(num[i], num[j]), (num[i] + num[j]));
          check(mul_opt(num[i], num[j]), (num[i] * num[j]));
          check(max_opt(num[i], num[j]), Math.max(num[i], num[j]));
          check(min_opt(num[i], num[j]), Math.min(num[i], num[j]));
        } 
      }
    }

    public void check(double a, double b) {
      if (a != b) {
        System.out.println("false");
        System.out.println(a);
        System.out.println(b);
        System.out.println();
      }
    }

    public double add_opt(double a, double b) throws Exception {
      return Math.max(a, b) + Math.min(a, b);
    }

    public double mul_opt(double a, double b) throws Exception {
      return Math.max(a, b) * Math.min(a, b);
    }

    public double max_opt(double a, double b) throws Exception {
      return Math.max(Math.max(a, b), Math.min(a, b));
    }

    public double min_opt(double a, double b) throws Exception {
      return Math.min(Math.max(a, b), Math.min(a, b));
    }
}
```
The `NaN` is a special case. Because `NaN == NaN` is false in Java, so I run the case and check the result. 

Should I add the other test cases for `NaN` ?